Update the input of GLPhotorealisticInstanceRender
