+++ examples/Cuda/Tutorials/GL_PhotorealisticInstanceRender/CMakeLists.txt
@@ -2,5 +2,5 @@ if(PERIDYNO_LIBRARY_PARTICLESYSTEM)
     set(LIB_DEPENDENCY
         Modeling
         GlfwGUI)
-    add_example(GL_InstancePhotorealisticRender Tutorials LIB_DEPENDENCY)
+    add_example(GL_PhotorealisticInstanceRender Tutorials LIB_DEPENDENCY)
 endif()
