+++ examples/Cuda/Tutorials/GL_PhotorealisticRender/main.cpp
