+++ src/Core/Backend/Cuda/Array/ArrayMap.inl
@@ -1,10 +1,84 @@
 #include "ArrayTools.h"
 #include "Algorithm/Scan.h"
 #include "Algorithm/Reduction.h"
-#include "Array/Array.h"
 
 namespace dyno
 {
+	template<class ElementType>
+	class ArrayMap<ElementType, DeviceType::GPU>
+	{
+	public:
+		ArrayMap()
+		{
+		};
+
+		/*!
+		*	\brief	Do not release memory here, call clear() explicitly.
+		*/
+		~ArrayMap() {};
+
+		/**
+		 * @brief Pre-allocate GPU space for
+		 *
+		 * @param counts
+		 * @return true
+		 * @return false
+		 */
+		bool resize(const DArray<uint> counts);
+		bool resize(const uint arraySize, const uint eleSize);
+
+		template<typename ET2>
+		bool resize(const ArrayMap<ET2, DeviceType::GPU>& src);
+
+		DYN_FUNC inline uint size() const { return m_maps.size(); }
+		DYN_FUNC inline uint elementSize() const { return m_elements.size(); }
+
+		GPU_FUNC inline Map<int, ElementType>& operator [] (unsigned int id) {
+			return m_maps[id];
+		}
+
+		GPU_FUNC inline const Map<int, ElementType>& operator [] (unsigned int id) const {
+			return m_maps[id];
+		}
+
+		DYN_FUNC inline bool isCPU() const { return false; }
+		DYN_FUNC inline bool isGPU() const { return true; }
+		DYN_FUNC inline bool isEmpty() const { return m_index.size() == 0; }
+
+		void clear();
+
+		void assign(const ArrayMap<ElementType, DeviceType::GPU>& src);
+		void assign(const ArrayMap<ElementType, DeviceType::CPU>& src);
+		void assign(std::vector<std::map<int, ElementType>>& src);
+
+		friend std::ostream& operator<<(std::ostream& out, const ArrayMap<ElementType, DeviceType::GPU>& aMap)
+		{
+			ArrayMap<ElementType, DeviceType::CPU> hMap;
+			hMap.assign(aMap);
+			out << hMap;
+
+			return out;
+		}
+
+		const DArray<uint>& index() const { return m_index; }
+		const DArray<Pair<int, ElementType>>& elements() const { return m_elements; }
+		const DArray<Map<int, ElementType>>& maps() const { return m_maps; }
+
+		/*!
+		*	\brief	To avoid erroneous shallow copy.
+		*/
+		ArrayMap<ElementType, DeviceType::GPU>& operator=(const ArrayMap<ElementType, DeviceType::GPU>&) = delete;
+
+	private:
+		DArray<uint> m_index;
+		DArray<Pair<int, ElementType>> m_elements;
+
+		DArray<Map<int, ElementType>> m_maps;
+	};
+
+	template<typename T>
+	using DArrayMap = ArrayMap<T, DeviceType::GPU>;
+
 	template<class ElementType>
 	void ArrayMap<ElementType, DeviceType::GPU>::clear()
 	{
